pyproject.toml
[project]
name = "civic-langgraph-migration"
version = "0.1.0"
description = "CIViC extraction pipeline migrated to LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # LangGraph & LangChain
    "langgraph>=0.2.0",
    "langgraph[sqlite]>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",

    # OpenAI-compatible client (for vLLM)
    "openai>=1.0.0",

    # Async HTTP
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",

    # PDF processing (same as original)
    "pymupdf>=1.24.0",
    "Pillow>=10.0.0",

    # Data handling
    "pydantic>=2.0.0",

    # Environment
    "python-dotenv>=1.0.0",

    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["graphs", "runtime", "config"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
